made a general user search functionality
